requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Deep Learning Frameworks
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter Notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Model Tracking & Experiment Management
mlflow>=2.7.0
tensorboard>=2.13.0

# Data Processing & Augmentation
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Natural Language Processing (if needed)
nltk>=3.8.0
spacy>=3.6.0

# Hyperparameter Tuning
optuna>=3.3.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
python-dotenv>=1.0.0
kagglehub>=0.2.0

# Model Serialization
h5py>=3.9.0

# Data Validation
pydantic>=2.0.0

# API & Deployment (optional)
fastapi>=0.103.0
uvicorn>=0.23.0

